python_agent/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
requests==2.31.0
gradio==4.44.1
python-dotenv==1.0.0
